NCD-672: Move appReducer.ts
