+++ src/index.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { App, render } from '@nordicsemiconductor/pc-nrfconnect-shared';
 
 import BoilerplateDeviceSelector from './app/DeviceSelector';
-import appReducer from './appReducer';
+import appReducer from './app/appReducer';
 import Configuration from './features/Configuration/Configuration';
 import SidePanel from './SidePanel';
